refactor(ci): move pre-commit hook commands into `toolr pre-commit`
Create a dedicated `pre-commit` command group under `tools/precommit/`
holding the commands that back this repo's prek hooks, completing the
dogfooding: `check-gate-needs` moves over from the `ci` group into
`tools/precommit/workflows.py`, and the standalone
`.pre-commit-hooks/regen-doc-snippets.py` script becomes
`toolr pre-commit regen-doc-snippets` in `tools/precommit/docs.py`,
folding in the step-summary remediation block the old
`toolr ci check-doc-snippets` wrapper wrote. Both the wrapper and
`.pre-commit-hooks/` are retired, and the regen command's subprocesses
now go through `ctx.run` (cwd/env passthrough; an empty stdin pipe keeps
clap honouring COLUMNS).

The docs job now dispatches the command via the Setup ToolR action: the
released binary only drives the dispatch — keeping the runner-schema
contract aligned with the released toolr-py in the tools venv — while
the freshly built binary under test remains what the captures execute,
via TOOLR_REGEN_BINARY. The regen command keeps the advisory no-op
behaviour locally (missing capture binary, or a dev venv that cannot
import `toolr._runner`) and hard-fails only when TOOLR_REGEN_PYTHON is
explicitly set but unusable.
