+++ mise.toml
@@ -22,8 +22,8 @@ rust = { version = "1.96.1", profile = "minimal", components = "clippy,rustfmt" 
 uv = { version = "0.11.28" }
 prek = "0.4.4"
 "cargo:cargo-edit" = "0.13.11"
-# Dogfood: the released toolr binary drives repo automation (the
-# `check-pipeline-gate-needs` prek hook runs `toolr ci check-gate-needs`).
+# Dogfood: the released toolr binary drives repo automation (the prek
+# hooks backed by `toolr pre-commit <command>` in `tools/precommit/`).
 # CI jobs install it via the in-repo Setup ToolR action instead, which
 # also verifies the SLSA attestation.
 "aqua:s0undt3ch/ToolR" = "0.26.0"
